Convert background page to event page
So Chrome can suspend it when not needed and safe resources
