+++ src/eventpage/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html>
 <head>
-    <title>Minimal Bookmarks Tree Background Page</title>
+    <title>Minimal Bookmarks Tree Event Page</title>
     <script src="index.js"></script>
 </head>
 <body>
