+++ src/manifest.json
@@ -6,7 +6,8 @@
 	"version": "4.5.0",
 	"minimum_chrome_version": "51",
 	"background": {
-		"page": "background/index.html"
+		"page": "eventpage/index.html",
+		"persistent": false
 	},
 	"omnibox": {
 		"keyword": "bm"
